+++ cmd/wehelp/main.go
@@ -23,7 +23,7 @@ func init() {
 		}
 	})
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default: ./wehelp.yaml or /etc/wehelp/wehelp.yaml)")
-	rootCmd.AddCommand(serveCmd, migrateCmd)
+	rootCmd.AddCommand(serveCmd, migrateCmd, seedCmd)
 }
 
 func main() {